+++ packages/tsurlfilter/rollup.config.ts
@@ -4,7 +4,6 @@ import globals from 'rollup-plugin-node-globals';
 import camelCase from 'lodash/camelCase';
 import typescript from '@rollup/plugin-typescript';
 import json from '@rollup/plugin-json';
-import nodePolyfills from 'rollup-plugin-polyfill-node';
 import cleanup from 'rollup-plugin-cleanup';
 import { terser } from 'rollup-plugin-terser';
 import { preserveShebangs } from 'rollup-plugin-preserve-shebangs';
@@ -32,7 +31,6 @@ const commonConfig = {
             sourceMap: false,
         }),
         globals(),
-        nodePolyfills(),
 
         // Allow node_modules resolution, so you can use 'external' to control
         // which external modules to include in the bundle
@@ -44,6 +42,17 @@ const commonConfig = {
         }),
     ],
 };
+
+const commonExternal = [
+    '@adguard/scriptlets',
+    'is-ip',
+    'punycode/',
+    'tldts',
+    'is-cidr',
+    'cidr-tools',
+    'zod',
+    'commander',
+];
 
 const esmConfig = {
     input: [
@@ -60,6 +69,7 @@ const esmConfig = {
             sourcemap: false,
         },
     ],
+    external: commonExternal,
     ...commonConfig,
 };
 
@@ -77,6 +87,7 @@ const esmDeclarativeConverterConfig = {
             sourcemap: false,
         },
     ],
+    external: commonExternal,
     ...commonConfig,
 };
 
@@ -115,25 +126,26 @@ const cliConfig = {
         },
     ],
     external: [
-        '@adguard/scriptlets',
-        'is-ip',
-        'punycode/',
-        'tldts',
-        'is-cidr',
-        'netmask',
-        'ip6addr',
-        'zod',
+        'fs',
         'path',
         'commander',
-        'fs',
     ],
-
     plugins: [
         // Allow json resolution
         json(),
 
         // Compile TypeScript files
         typescript(),
+
+        // Allow bundling cjs modules (unlike webpack, rollup doesn't understand cjs)
+        commonjs({
+            sourceMap: false,
+        }),
+
+        // Allow node_modules resolution, so you can use 'external' to control
+        // which external modules to include in the bundle
+        // https://github.com/rollup/rollup-plugin-node-resolve#usage
+        resolve({ preferBuiltins: false }),
 
         cleanup({
             comments: ['srcmaps'],